Persist monotonic scene identity high-water marks

+++ src/main/java/games/pixscape/runtime/api/EntityRef.java
@@ -15,7 +15,7 @@ public interface EntityRef {
     /**
      * Preferred persistent/public entity identity.
      */
-    long stableId();
+    int stableId();
 
     boolean exists();
